improve configuration
