+++ centreon/packages/js-config/package.json
@@ -18,10 +18,13 @@
   "devDependencies": {
     "@badeball/cypress-cucumber-preprocessor": "^18.0.5",
     "@bahmutov/cypress-esbuild-preprocessor": "^2.2.0",
+    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
+    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
     "@tsconfig/node16": "^1.0.4",
     "@types/cypress-cucumber-preprocessor": "^4.0.2",
     "@types/dockerode": "^3.3.19",
     "dockerode": "^3.3.5",
+    "esbuild": "^0.19.2",
     "eslint": "^8.17.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-prettier": "^8.5.0",
