+++ centreon/tests/e2e/cypress/e2e/Platform-upgrade-update/02-platform-upgrade/index.ts
@@ -93,7 +93,7 @@ Given(
           }
         ]
       }).then(() => {
-        Cypress.config('baseUrl', 'http://0.0.0.0:4000');
+        Cypress.config('baseUrl', 'http://127.0.0.1:4000');
 
         return cy
           .intercept('/waiting-page', {
